Throw out web ui to accomodate for requirements change
The objectives for the WAT2023 project changed and the browser interface is no longer
required to complete the project. Because of that mostly the docker configuration was
changed but it is possible that some changes from the previous work which was focused
on the ui and was not commited yet slipped into this commit.