close #49
Besides untracking the .env files the docker-compose.yml got modified
to rely on .env file only. This way a developer is not required to have
two .env files. This is meant to limit the mental overhead required to
start with the project but it has a downside of causing confilcts/errors
while alternating between deployment types (local vs docker).

Additionally the CONTRIBUTING.md got updated. Finally no more notes left
from the old definition of the project.